NET-171: move info to rootdomain extractor file
